fix: better async tests and fix invoke race cond
There were situations where an invoked task could still execute after
the machine has transitioned/exited. This is now handled with
:erlang.alias/0, which creates an alias for task to send events to that
can be de-aliased when the task is canceled. This prevents a case where
the task is canceled and the interpreter stopped, but the task still
sends the event, causing an error because the process is no longer
alive.