change!: major change to events and event matching
Events are no longer of the form `{"string", data}`, but are instead
arbitrary terms that can be matched on. This means machine
configs are now defined with a `defmachine` macro so that we can
auto-generate pattern-matching functions.

This is part of an effort to make Protean more general-purpose and
useful in a larger variety of contexts, e.g. responding to events that
are defined as structs (`%MyEvent{}`) or non-string tuples
(`{MyNamespace, %MyEvent{}}`).