change!: subscribe directly to machine, not id
This simplifies subscriptions and the return value of `start_machine/2`
by hiding the UUID that we use internally as a `Phoenix.PubSub` topic.
Instead of needing to track the machine pid and the id to subscribe, you
can subscribe using the machine pid, and get `{:ok, id}` as a return
value so that you can reference `id` if you need to identify the source
of the message sent to you.